Phase 1: magic bitboard attack tables, exhaustively verified
Generate all 128 magics offline (tools/gen_magics.py, seed 0xC0FFEE) and embed them so the
engine pays no search cost at init — only a jitted table fill. Rook table is 102,400 entries,
bishop 5,248, 841 KB total; engine.bitboard import costs 3.01s of the 90s budget (4.76s
including numpy+numba).

engine/bitboard.py deliberately does not self-verify. Verification runs in
tools/verify_movegen.py instead, over *every* occupancy subset of every square rather than a
sample: a bad magic collides on some rare occupancy and produces an illegal move once every
few thousand positions, which is exactly the failure a sample misses. All 107,648 subsets
check out against a reference ray-walker.

Also extends mypy from ["agent.py", "harness"] to cover engine/ and tools/ — the earlier
"gate is clean" was true but vacuous for every file written this session.

+++ tools/sprt.py
@@ -83,7 +83,7 @@ def opening_fens() -> list[str]:
 
 
 def elo_to_score(elo: float) -> float:
-    return 1.0 / (1.0 + 10.0 ** (-elo / 400.0))
+    return float(1.0 / (1.0 + 10.0 ** (-elo / 400.0)))
 
 
 def score_to_elo(score: float) -> float:
